+++ src/navigation/AlarmsStack.jsx
@@ -2,6 +2,7 @@ import React from "react";
 import { createNativeStackNavigator } from "@react-navigation/native-stack";
 import { screen } from "../constant/screenName";
 import { AlarmsScreen } from "../screens/AlarmsScreen";
+import { ProgramAlarmsForm } from "../screens/ProgramAlarmsForm/ProgramAlarmsForm";
 
 const Stack = createNativeStackNavigator();
 
@@ -17,6 +18,13 @@ export function AlarmsStack() {
         name={screen.alarms.alarms} // Es el nombre que se va a usar para navegar
         component={AlarmsScreen} // Es el componente que se va a renderizar
         options={{ title: "Programar", headerTitleAlign: "center" }} // Es el nombre que aparece en el header
+      />
+
+      
+      <Stack.Screen
+        name={screen.alarms.programAlarms}
+        component={ProgramAlarmsForm} //Estableciendo redireccion a ProgramAlarmsForm
+        options={{ title: "Programar alarma", headerTitleAlign: "center" }}
       />
     </Stack.Navigator>
   );
