+++ src/screens/AlarmsScreen.jsx
@@ -1,9 +1,9 @@
 import { View, Text } from "react-native";
 import React from "react";
-import {ProgramAlarms} from '../components/ProgramAlarms/ProgramAlarms';
+import {Alarms} from '../components/Alarms/Alarms';
 
 export function AlarmsScreen() {
   return (
-    <ProgramAlarms/>
+    <Alarms/>
   );
 }
